\name{Cmeasure}
\alias{Cmeasure}
\title{C-Measure subtypes}
\arguments{
\item{x}{Vektor of points in input space}
\item{y}{Vektor of points in outputsapce }
\item{k}{number of nearest neighbors, both measures set it always to k=1}
\value{[[1:2] numerical vector of MinimalPathlength and MinimWiring values}
\description{
  Calculate the C-Measure subtypes of minimal path length and minimal wirin}
\author{Michael Thrun}
