\name{SpearmansRho}
\alias{SpearmansRho}
\title{Calculates the error of a projection with spearman's rank correlation coefficient}
\usage{SpearmansRho(InputDists, OutputDists)}
\arguments{
\item{InputDists}{[1:d,1:d] numeric matrix with input distances}
\item{OutputDists}{[1:d,1:d] numeric matrix with output distances}
}
\value{rho}
\description{Calculates the error of a projection with spearman's rank
correlation coefficient}
\author{Julian Märte}
