\name{Cmeasure}
\alias{Cmeasure}
\title{C-Measure subtypes}
\description{Calculate the C-Measure subtypes of minimal path length and minimal
wiring}
\arguments{
\item{x}{Vector of points in input space.}
\item{y}{Vector of points in output space.}
\item{k}{Number of nearest neighbors, both measures set it always to k=1.}
}
\value{[[1:2] Numerical vector of MinimalPathlength and MinimalWiring values.}
\author{Michael Thrun}
